feat(actionable_guidance): standardize actionable guidance for startup errors
- Simplify and standardize startup error guidance messages with a unified template.
- Replace shell_quote with powershell_quote for Windows PowerShell quoting.
- Render precise next command suggestions based on configured daemon binary path or name.
- Move actionable guidance tests to dedicated unit test file for better organization.
- Improve guidance for LifecycleError variants: LaunchDaemon, StartupFailed, StartupTimeout, and StartupAborted.
- Enhance unknown operation rendering to handle cases with no known operations by suggesting domain help.

This improves operator experience by consistently presenting helpful, actionable advice on startup failures and command usage.
