+++ crates/weaver-cli/src/tests/unit.rs
@@ -372,6 +372,7 @@ fn is_daemon_not_running_rejects_non_connect_errors() {
     let ser_err = AppError::SerialiseRequest(serde_json::from_str::<()>("bad").unwrap_err());
     assert!(!is_daemon_not_running(&ser_err));
 }
+mod actionable_guidance;
 mod after_help;
 mod auto_start;
 mod bare_invocation;
